modify sql

+++ db/punch-4.0-delta-schema.sql
@@ -40,7 +40,7 @@ CREATE TABLE `eh_punch_vacation_balance_logs` (
 ALTER TABLE eh_punch_statistics ADD COLUMN `annual_leave_balance` DOUBLE COMMENT '年假余额';
 ALTER TABLE eh_punch_statistics ADD COLUMN `overtime_compensation_balance` DOUBLE COMMENT '调休余额';
 ALTER TABLE eh_punch_statistics ADD COLUMN `device_change_counts` INT COMMENT '设备异常次数';
-ALTER TABLE eh_punch_statistics ADD COLUMN `exception_request_counts` DOUBLE COMMENT '异常申报次数';
+ALTER TABLE eh_punch_statistics ADD COLUMN `exception_request_counts` INT COMMENT '异常申报次数';
 ALTER TABLE eh_punch_statistics ADD COLUMN `belate_time` DOUBLE COMMENT '迟到时长(小时)';
 ALTER TABLE eh_punch_statistics ADD COLUMN `leave_early_time` DOUBLE COMMENT '早退时长(小时)';
 ALTER TABLE eh_punch_statistics ADD COLUMN `forgot_count` INT COMMENT '下班缺卡次数';
